add 4 logs
addtagfile deletetagfile restorefile updatefiletitle

+++ back/files/deleteTag.php
@@ -1,8 +1,19 @@
 <?php
 include("../database.php");
+session_start();
 
 $IDFichier = $_POST['IDFichier'];
 $IDTag = $_POST['IDTag'];
+
+
+//Get the name of the tag
+$tagName = "SELECT NomTag FROM tag WHERE IDTag = '$IDTag'";
+$result = $mysqli->query($tagName);
+
+//Check for errors
+if (!$result) {
+    die("Erreur lors de la récupération du nom du tag pour logs : " . $mysqli->error);
+}
 
 // Suppression du lien entre le fichier et le tag
 $sql = "DELETE FROM `classifier` WHERE `classifier`.`IDFichier` = '$IDFichier' AND `classifier`.`IDTag` = '$IDTag'";
@@ -32,5 +43,18 @@ if ($result->num_rows == 0) {
         die('Erreur d\'ajout du tag \'Sans tag\' sur un fichier sans tag : ' . $mysqli->error);
     }
 }
+
+//Get the name of the file
+$fileName = "SELECT Nom FROM fichier WHERE IDFichier = '$IDFichier'";
+$result = $mysqli->query($fileName);
+
+//Check for errors
+if (!$result) {
+    die("Erreur lors de la récupération du nom du fichier pour logs : " . $mysqli->error);
+}
+
+// INSERT LOG
+include '../logRegister.php';
+registerNewLog($mysqli, $_SESSION['id'], "Retire le tag : " . $tagName . " au fichier : " . $fileName);
 
 echo "OK";
